pyproject.toml
[tool.poetry]
name = "pydantic-langchain-demo"
version = "0.1.0"
description = "A multi-agent pipeline for researching topics and generating social media content"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/yourusername/pydantic-langchain-demo"
packages = [{include = "models"}, {include = "agents"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
pydantic = "^2.0.0"
openai = "^1.0.0"
langgraph = "^0.2.3"
python-dotenv = "^1.0.0"
streamlit = "^1.28.0"
asyncio = "^3.4.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.4.0"

[tool.poetry.scripts]
app = "app:main"
cli = "cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
